requirements.txt
openpyxl==3.1.5
google-generativeai==0.8.3
pandas==2.2.3
pillow==11.0.0
pygame==2.6.1
python-dotenv==1.0.1